requirements.txt
flask
flask-pymongo
flask-jwt-extended
werkzeug
python-dotenv
